pyproject.toml
[project]
name = "moondream-multimodal-analysis"
version = "0.1.0"
description = "A Streamlit app for multimodal visual language model analysis using Moondream and Gemini"
authors = [
    {name = "William VanBuskirk", email = "william.n.vanbuskirk@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.8"
dependencies = [
    "streamlit>=1.28.0",
    "pillow>=10.0.0",
    "opencv-python>=4.8.0",
    "numpy>=1.24.0",
    "streamlit-webrtc>=0.47.0",
    "google-genai>=0.1.0",
    "requests>=2.31.0",
    "moondream>=0.1.1",
]

[project.optional-dependencies]
dev = [
    "black>=23.0.0",
    "flake8>=6.0.0",
    "pytest>=7.4.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.black]
line-length = 88
target-version = ['py38']

[tool.flake8]
max-line-length = 88
extend-ignore = ["E203", "W503"] 